# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Affilist API
  description: |
    REST API for the Affiliate Link Aggregator application.
    
    This API provides endpoints for managing affiliate links, categories, analytics, and admin operations.
    
    ## Authentication
    
    Admin endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    API requests are rate limited to prevent abuse:
    - Public endpoints: 100 requests per minute per IP
    - Admin endpoints: 1000 requests per minute per authenticated user
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns error details in JSON format:
    ```json
    {
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: Affilist Support
    email: support@affilist.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api-staging.affilist.com/api
    description: Staging server
  - url: https://api.affilist.com/api
    description: Production server
paths:
...
  /clicks:
    post:
      summary: Track affiliate link click
      description: Record a click event for an affiliate link
      tags:
        - Analytics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - linkId
              properties:
                linkId:
                  type: string
                  description: ID of the clicked affiliate link
                referrer:
                  type: string
                  description: Referrer URL
                userAgent:
                  type: string
                  description: User agent string
      responses:
        '201':
          description: Click recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  clickId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /redirect/{linkId}:
    get:
      summary: Redirect to affiliate link
      description: Track click and redirect to the affiliate URL
      tags:
        - Analytics
      parameters:
        - name: linkId
          in: path
          required: true
          description: Affiliate link ID
          schema:
            type: string
      responses:
        '302':
          description: Redirect to affiliate URL
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              field: "email"
              message: "Invalid email format"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
...